frontend: chats: Move header to ChatView
