+++ frontend/srcs/src/pages/chat/chat.tsx
@@ -1,30 +1,9 @@
 import { RouteComponentProps } from "react-router";
-import { NavLink } from "react-router-dom";
-import { ChatHeader } from "../../components/chat/ChatHeader";
 
 import { ChatNavBar } from "../../components/chat/ChatNavBar";
+import { ChatView } from "../../components/chat/ChatView";
 import { Chat, ChatType } from "../../models/Chat";
 //import ChannelSearch from "./channelSearch";
-
-type TooltipIconButtonProps = {
-    tooltip: string;
-    icon: string;
-    iconStyle: string;
-    href: string;
-}
-
-export function TooltipIconButton({ icon, iconStyle, tooltip, href }: TooltipIconButtonProps) {
-    return (
-        <NavLink className="flex flex-col items-center has-tooltip" activeClassName="text-blue-600" to={href}>
-            <i className={`${iconStyle} ${icon}`}/>
-            <span className="p-1 mt-6 bg-gray-100 rounded shadow-lg tooltip">{tooltip}</span>
-        </NavLink>
-    );
-}
-
-TooltipIconButton.defaultProps = {
-    iconStyle: "fas fa-lg"
-}
 
 type ChatPageParams = {
     id: string;
@@ -39,8 +18,6 @@ export default function ChatPage({ match }: RouteComponentProps<ChatPageParams>)
 
     console.log("Current chat " + match.params.id);
 
-    
-
     return (<div className="flex h-full">
         {/*<Route exact path='/chat/find'>
             <ChannelSearch
@@ -53,10 +30,6 @@ export default function ChatPage({ match }: RouteComponentProps<ChatPageParams>)
         </Route>
         */}
         <ChatNavBar chats={chats} activeIndex={Number(match.params.id)}></ChatNavBar>
-        <ChatHeader chat={currentChat}>
-            <span>40 users</span>
-            <TooltipIconButton tooltip="Invite" icon="fa-user-plus" href={`/chat/${chatId}/settings`}></TooltipIconButton>
-            <TooltipIconButton tooltip="Settings" icon="fa-cog" href={`/chat/${chatId}/settings`}></TooltipIconButton>
-        </ChatHeader>
+        <ChatView className="flex-1" chat={currentChat}></ChatView>
     </div>)
 }
